Thiet ke thanh toan Redeem Reward cho user

+++ back-end/src/main/java/nlu/com/app/dto/json/RedeemRewardJson.java
@@ -22,7 +22,7 @@ public class RedeemRewardJson {
     String material;
     String size;
     String weight;
-    long qty_in_stock;
+    int qty_in_stock;
     String [] images;
     String url;
